+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Ensure BigDecimal.new always receives a string - compatibility fix for ruby-1.9.2-p320
 - Update Maldivian Currency to MVR and fix ރ. to be ރ
+- Add exponent to currency
 
 ## 5.1.0